streetlight/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.46
Flask-WTF==1.0.0
WTForms==3.0.0
Flask-Login==0.5.0
python-dotenv==0.19.1
cryptography==36.0.0
email-validator==1.1.3
Werkzeug==2.0.1
